rename : coverLetter/new 구조 조정
